move save_images_with_boxes
